Cleaned up stuff for bash_profile. Updated git-aware-prompt upstream. Improved install.sh slightly.

+++ install.sh
@@ -1,9 +1,11 @@
 cd ~/
 
 # Make a backup, hey you never know...
-mv ~/.bash_profile ~/.dotfiles/.bash_profile.bak
+if [ -f ~/.bash_profile ]; then
+  mv ~/.bash_profile ~/.dotfiles/.bash_profile.bak
+fi
 
-ln -s ~/.dotfiles/.bash_profile .bash_profile
+ln -s ~/.dotfiles/.bash_profile ~/.bash_profile
 
 # want to source it, but getting error
 # source ~/.bash_profile
